requirements.txt
cachetools==3.1.1
certifi==2018.11.29
chardet==3.0.4
configparser==3.8.1
DateTimeRange==0.6.1
EasyProcess==0.2.5
ez-setup==0.9
file-cache==0.1.43
google-api-python-client==1.7.11
google-auth==1.7.0
google-auth-httplib2==0.0.3
httplib2==0.14.0
idna==2.8
mbstrdecoder==0.8.1
oauth2client==4.1.3
parse==1.9.0
parse-type==0.4.2
pyasn1==0.4.7
pyasn1-modules==0.2.7
python-dateutil==2.8.1
pytz==2019.3
PyVirtualDisplay==0.2.1
PyYAML==5.1.2
requests==2.21.0
rsa==4.0
selenium==3.141.0
six==1.12.0
typepy==0.6.0
uritemplate==3.0.0
urllib3==1.24.1
